+++ code/src/cljs/sixsq/nuvla/ui/pricing/spec.cljs
@@ -1,0 +1,9 @@
+(ns sixsq.nuvla.ui.pricing.spec
+  (:require
+    [clojure.spec.alpha :as s]))
+
+
+(s/def ::db (s/keys :req []))
+
+
+(def defaults {})